+++ android/expoview/src/main/Common/cpp/SharedItems/RemoteObject.cpp
@@ -1,13 +1,13 @@
 #include "RemoteObject.h"
 #include "SharedParent.h"
-#include "NativeReanimatedModule.h"
+#include "RuntimeDecorator.h"
 #include <jsi/jsi.h>
 
 using namespace facebook;
 
 namespace reanimated {
 
-void RemoteObject::maybeInitializeOnUIRuntime(jsi::Runtime &rt) {
+void RemoteObject::maybeInitializeOnWorkletRuntime(jsi::Runtime &rt) {
   if (initializer.get() != nullptr) {
     backing = getWeakRef(rt);
     *backing.lock() = initializer->shallowClone(rt);
@@ -16,14 +16,14 @@ void RemoteObject::maybeInitializeOnUIRuntime(jsi::Runtime &rt) {
 }
 
 jsi::Value RemoteObject::get(jsi::Runtime &rt, const jsi::PropNameID &name) {
-  if (module->isUIRuntime(rt)) {
+  if (RuntimeDecorator::isWorkletRuntime(rt)) {
     return backing.lock()->getObject(rt).getProperty(rt, name);
   }
   return jsi::Value::undefined();
 }
 
 void RemoteObject::set(jsi::Runtime &rt, const jsi::PropNameID &name, const jsi::Value &value) {
-  if (module->isUIRuntime(rt)) {
+  if (RuntimeDecorator::isWorkletRuntime(rt)) {
     backing.lock()->getObject(rt).setProperty(rt, name, value);
   }
   // TODO: we should throw if trying to update remote from host runtime