+++ docs/pages/_app.tsx
@@ -8,7 +8,7 @@ import React, { useState, useEffect } from 'react';
 import { TrackPageView } from '~/common/analytics';
 import { preprocessSentryError } from '~/common/sentry-utilities';
 import 'react-diff-view/style/index.css';
-import '@expo/styleguide/dist/expo-colors.css';
+import '@expo/styleguide/dist/expo-theme.css';
 import 'tippy.js/dist/tippy.css';
 import '../public/static/libs/algolia/algolia.css';
 import '../public/static/libs/algolia/algolia-mobile.css';