+++ ios/Exponent/Versioned/Core/Api/Components/Svg/Elements/RNSVGGroup.m
@@ -38,7 +38,7 @@
 
     __block CGRect bounds = CGRectNull;
 
-    [self traverseSubviews:^(UIView *node) {
+    [self traverseSubviews:^(RNSVGView *node) {
         if ([node isKindOfClass:[RNSVGMask class]] || [node isKindOfClass:[RNSVGClipPath class]]) {
             // no-op
         } else if ([node isKindOfClass:[RNSVGNode class]]) {
@@ -160,7 +160,7 @@
     return cached;
 }
 
-- (UIView *)hitTest:(CGPoint)point withEvent:(UIEvent *)event
+- (RNSVGPlatformView *)hitTest:(CGPoint)point withEvent:(UIEvent *)event
 {
     CGPoint transformed = CGPointApplyAffineTransform(point, self.invmatrix);
     transformed = CGPointApplyAffineTransform(transformed, self.invTransform);
@@ -192,7 +192,7 @@
         }
     }
 
-    for (UIView *node in [self.subviews reverseObjectEnumerator]) {
+    for (RNSVGView *node in [self.subviews reverseObjectEnumerator]) {
         if ([node isKindOfClass:[RNSVGNode class]]) {
             if ([node isKindOfClass:[RNSVGMask class]]) {
                 continue;
@@ -201,21 +201,21 @@
             if (event) {
                 svgNode.active = NO;
             }
-            UIView *hitChild = [svgNode hitTest:transformed withEvent:event];
+            RNSVGPlatformView *hitChild = [svgNode hitTest:transformed withEvent:event];
             if (hitChild) {
                 svgNode.active = YES;
                 return (svgNode.responsible || (svgNode != hitChild)) ? hitChild : self;
             }
         } else if ([node isKindOfClass:[RNSVGSvgView class]]) {
             RNSVGSvgView* svgView = (RNSVGSvgView*)node;
-            UIView *hitChild = [svgView hitTest:transformed withEvent:event];
+            RNSVGPlatformView *hitChild = [svgView hitTest:transformed withEvent:event];
             if (hitChild) {
                 return hitChild;
             }
         }
     }
 
-    UIView *hitSelf = [super hitTest:transformed withEvent:event];
+    RNSVGPlatformView *hitSelf = [super hitTest:transformed withEvent:event];
     if (hitSelf) {
         return hitSelf;
     }