+++ ios/vendored/unversioned/react-native-screens/ios/RNSScreen.h
@@ -21,6 +21,8 @@ typedef NS_ENUM(NSInteger, RNSScreenStackAnimation) {
   RNSScreenStackAnimationNone,
   RNSScreenStackAnimationFade,
   RNSScreenStackAnimationFlip,
+  RNSScreenStackAnimationSlideFromBottom,
+  RNSScreenStackAnimationSimplePush,
 };
 
 typedef NS_ENUM(NSInteger, RNSScreenReplaceAnimation) {
@@ -34,10 +36,29 @@ typedef NS_ENUM(NSInteger, RNSActivityState) {
   RNSActivityStateOnTop = 2
 };
 
+typedef NS_ENUM(NSInteger, RNSStatusBarStyle) {
+  RNSStatusBarStyleAuto,
+  RNSStatusBarStyleInverted,
+  RNSStatusBarStyleLight,
+  RNSStatusBarStyleDark,
+};
+
+typedef NS_ENUM(NSInteger, RNSWindowTrait) {
+  RNSWindowTraitStyle,
+  RNSWindowTraitAnimation,
+  RNSWindowTraitHidden,
+  RNSWindowTraitOrientation,
+};
+
 @interface RCTConvert (RNSScreen)
 
 + (RNSScreenStackPresentation)RNSScreenStackPresentation:(id)json;
 + (RNSScreenStackAnimation)RNSScreenStackAnimation:(id)json;
+
+#if !TARGET_OS_TV
++ (RNSStatusBarStyle)RNSStatusBarStyle:(id)json;
++ (UIInterfaceOrientationMask)UIInterfaceOrientationMask:(id)json;
+#endif
 
 @end
 
@@ -67,6 +88,17 @@ typedef NS_ENUM(NSInteger, RNSActivityState) {
 @property (nonatomic) RNSScreenStackAnimation stackAnimation;
 @property (nonatomic) RNSScreenStackPresentation stackPresentation;
 @property (nonatomic) RNSScreenReplaceAnimation replaceAnimation;
+@property (nonatomic) BOOL hasOrientationSet;
+@property (nonatomic) BOOL hasStatusBarStyleSet;
+@property (nonatomic) BOOL hasStatusBarAnimationSet;
+@property (nonatomic) BOOL hasStatusBarHiddenSet;
+
+#if !TARGET_OS_TV
+@property (nonatomic) RNSStatusBarStyle statusBarStyle;
+@property (nonatomic) UIStatusBarAnimation statusBarAnimation;
+@property (nonatomic) BOOL statusBarHidden;
+@property (nonatomic) UIInterfaceOrientationMask screenOrientation;
+#endif
 
 - (void)notifyFinishTransitioning;