+++ ios/vendored/unversioned/react-native-screens/ios/RNSScreenStackHeaderConfig.h
@@ -2,13 +2,7 @@
 #import <React/RCTConvert.h>
 
 #import "RNSScreen.h"
-
-typedef NS_ENUM(NSInteger, RNSStatusBarStyle) {
-  RNSStatusBarStyleAuto,
-  RNSStatusBarStyleInverted,
-  RNSStatusBarStyleLight,
-  RNSStatusBarStyleDark,
-};
+#import "RNSSearchBar.h"
 
 @interface RNSScreenStackHeaderConfig : UIView
 
@@ -35,26 +29,12 @@ typedef NS_ENUM(NSInteger, RNSStatusBarStyle) {
 @property (nonatomic, retain) UIColor *largeTitleColor;
 @property (nonatomic) BOOL hideBackButton;
 @property (nonatomic) BOOL backButtonInCustomView;
+@property (nonatomic) BOOL disableBackButtonMenu;
 @property (nonatomic) BOOL hideShadow;
 @property (nonatomic) BOOL translucent;
 @property (nonatomic) UISemanticContentAttribute direction;
 
-#if !TARGET_OS_TV
-@property (nonatomic) RNSStatusBarStyle statusBarStyle;
-@property (nonatomic) UIStatusBarAnimation statusBarAnimation;
-@property (nonatomic) BOOL statusBarHidden;
-@property (nonatomic) UIInterfaceOrientationMask screenOrientation;
-#endif
-
 + (void)willShowViewController:(UIViewController *)vc animated:(BOOL)animated withConfig:(RNSScreenStackHeaderConfig*)config;
-+ (void)updateWindowTraits;
-
-#if !TARGET_OS_TV
-+ (UIStatusBarStyle)statusBarStyleForRNSStatusBarStyle:(RNSStatusBarStyle)statusBarStyle;
-+ (UIInterfaceOrientation)defaultOrientationForOrientationMask:(UIInterfaceOrientationMask)orientationMask;
-+ (UIInterfaceOrientation)interfaceOrientationFromDeviceOrientation:(UIDeviceOrientation)deviceOrientation;
-+ (UIInterfaceOrientationMask)maskFromOrientation:(UIInterfaceOrientation)orientation;
-#endif
 
 @end
 
@@ -68,6 +48,7 @@ typedef NS_ENUM(NSInteger, RNSScreenStackHeaderSubviewType) {
   RNSScreenStackHeaderSubviewTypeRight,
   RNSScreenStackHeaderSubviewTypeTitle,
   RNSScreenStackHeaderSubviewTypeCenter,
+  RNSScreenStackHeaderSubviewTypeSearchBar,
 };
 
 @interface RCTConvert (RNSScreenStackHeader)
@@ -75,11 +56,6 @@ typedef NS_ENUM(NSInteger, RNSScreenStackHeaderSubviewType) {
 + (RNSScreenStackHeaderSubviewType)RNSScreenStackHeaderSubviewType:(id)json;
 + (UIBlurEffectStyle)UIBlurEffectStyle:(id)json;
 + (UISemanticContentAttribute)UISemanticContentAttribute:(id)json;
-
-#if !TARGET_OS_TV
-+ (RNSStatusBarStyle)RNSStatusBarStyle:(id)json;
-+ (UIInterfaceOrientationMask)UIInterfaceOrientationMask:(id)json;
-#endif
 
 @end