+++ ios/vendored/unversioned/react-native-screens/ios/RNSScreenStackHeaderConfig.m
@@ -1,5 +1,6 @@
 #import "RNSScreenStackHeaderConfig.h"
 #import "RNSScreen.h"
+#import "RNSSearchBar.h"
 
 #import <React/RCTBridge.h>
 #import <React/RCTUIManager.h>
@@ -40,6 +41,39 @@
   }
   return self;
 }
+
+- (void) reactSetFrame:(CGRect)frame
+{
+  // Block any attempt to set coordinates on RNSScreenStackHeaderSubview. This
+  // makes UINavigationBar the only one to control the position of header content.
+  [super reactSetFrame:CGRectMake(0, 0, frame.size.width, frame.size.height)];
+}
+
+@end
+
+@interface RNSUIBarButtonItem : UIBarButtonItem
+
+@property (nonatomic) BOOL menuHidden;
+
+@end
+
+@implementation RNSUIBarButtonItem
+
+- (void)setMenuHidden:(BOOL)menuHidden
+{
+  _menuHidden = menuHidden;
+}
+
+#if defined(__IPHONE_OS_VERSION_MAX_ALLOWED) && defined(__IPHONE_14_0) && \
+    __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_14_0
+- (void)setMenu:(UIMenu *)menu {
+  if (@available(iOS 14.0, *)) {
+    if (!_menuHidden) {
+        super.menu = menu;
+    }
+  }
+}
+#endif
 
 @end
 
@@ -374,21 +408,6 @@ API_AVAILABLE(ios(13.0)){
 
   [navctr setNavigationBarHidden:shouldHide animated:animated];
 
-#if !TARGET_OS_TV
-  // we put it before check with return because we want to apply changes to status bar even if the header is hidden
-  if (config != nil) {
-    if (config.statusBarStyle || config.statusBarAnimation || config.statusBarHidden) {
-      [RNSScreenStackHeaderConfig assertViewControllerBasedStatusBarAppearenceSet];
-      if ([vc isKindOfClass:[RNSScreen class]]) {
-        [RNSScreenStackHeaderConfig updateStatusBarAppearance];
-      }
-    }
-    if (config.screenOrientation) {
-      [RNSScreenStackHeaderConfig enforceDesiredDeviceOrientation];
-    }
-  }
-#endif
-
   if (config.direction == UISemanticContentAttributeForceLeftToRight || config.direction == UISemanticContentAttributeForceRightToLeft) {
     navctr.view.semanticContentAttribute = config.direction;
     navctr.navigationBar.semanticContentAttribute = config.direction;
@@ -400,12 +419,16 @@ API_AVAILABLE(ios(13.0)){
   
   navitem.title = config.title;
 #if !TARGET_OS_TV
-  if (config.backTitle != nil || config.backTitleFontFamily || config.backTitleFontSize) {
-    prevItem.backBarButtonItem = [[UIBarButtonItem alloc]
-                                  initWithTitle:config.backTitle ?: prevItem.title
-                                  style:UIBarButtonItemStylePlain
-                                  target:nil
-                                  action:nil];
+  if (config.backTitle != nil || config.backTitleFontFamily || config.backTitleFontSize || config.disableBackButtonMenu) {
+    RNSUIBarButtonItem *backBarButtonItem = [[RNSUIBarButtonItem alloc]
+                                                 initWithTitle:config.backTitle ?: prevItem.title
+                                                 style:UIBarButtonItemStylePlain
+                                                 target:nil
+                                                 action:nil];
+    
+    [backBarButtonItem setMenuHidden:config.disableBackButtonMenu];
+    
+    prevItem.backBarButtonItem = backBarButtonItem;
     if (config.backTitleFontFamily || config.backTitleFontSize) {
       NSMutableDictionary *attrs = [NSMutableDictionary new];
       NSNumber *size = config.backTitleFontSize ?: @17;
@@ -485,6 +508,15 @@ API_AVAILABLE(ios(13.0)){
         navitem.titleView = subview;
         break;
       }
+      case RNSScreenStackHeaderSubviewTypeSearchBar: {
+        if ([subview.subviews[0] isKindOfClass:[RNSSearchBar class]]) {
+          if (@available(iOS 11.0, *)) {
+            RNSSearchBar *searchBar = subview.subviews[0];
+            navitem.searchController = searchBar.controller;
+            navitem.hidesSearchBarWhenScrolling = searchBar.hideWhenScrolling;
+          }
+        }
+      }
       case RNSScreenStackHeaderSubviewTypeBackButton: {
         break;
       }
@@ -522,183 +554,6 @@ API_AVAILABLE(ios(13.0)){
   }
 }
 
-#if !TARGET_OS_TV
-+ (void)assertViewControllerBasedStatusBarAppearenceSet
-{
-  static dispatch_once_t once;
-  static bool viewControllerBasedAppearence;
-  dispatch_once(&once, ^{
-    viewControllerBasedAppearence = [[[NSBundle mainBundle] objectForInfoDictionaryKey:@"UIViewControllerBasedStatusBarAppearance"] boolValue];
-  });
-  if (!viewControllerBasedAppearence) {
-    RCTLogError(@"If you want to change the appearance of status bar, you have to change \
-    UIViewControllerBasedStatusBarAppearance key in the Info.plist to YES");
-  }
-}
-#endif
-
-+ (void)updateStatusBarAppearance
-{
-#if !TARGET_OS_TV
-  [UIView animateWithDuration:0.4 animations:^{ // duration based on "Programming iOS 13" p. 311 implementation
-#if defined(__IPHONE_OS_VERSION_MAX_ALLOWED) && defined(__IPHONE_13_0) && \
-    __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_13_0
-    if (@available(iOS 13, *)) {
-      UIWindow *firstWindow = [[[UIApplication sharedApplication] windows] firstObject];
-      if (firstWindow != nil) {
-        [[firstWindow rootViewController] setNeedsStatusBarAppearanceUpdate];
-      }
-    } else
-#endif
-    {
-      [UIApplication.sharedApplication.keyWindow.rootViewController setNeedsStatusBarAppearanceUpdate];
-    }
-  }];
-#endif
-}
-
-#if !TARGET_OS_TV
-+ (UIStatusBarStyle)statusBarStyleForRNSStatusBarStyle:(RNSStatusBarStyle)statusBarStyle
-{
-#if defined(__IPHONE_OS_VERSION_MAX_ALLOWED) && defined(__IPHONE_13_0) && \
-    __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_13_0
-  if (@available(iOS 13.0, *)) {
-    switch (statusBarStyle) {
-      case RNSStatusBarStyleAuto:
-          return [UITraitCollection.currentTraitCollection userInterfaceStyle] == UIUserInterfaceStyleDark ? UIStatusBarStyleLightContent : UIStatusBarStyleDarkContent;
-      case RNSStatusBarStyleInverted:
-          return [UITraitCollection.currentTraitCollection userInterfaceStyle] == UIUserInterfaceStyleDark ? UIStatusBarStyleDarkContent : UIStatusBarStyleLightContent;
-      case RNSStatusBarStyleLight:
-          return UIStatusBarStyleLightContent;
-      case RNSStatusBarStyleDark:
-          return UIStatusBarStyleDarkContent;
-      default:
-        return UIStatusBarStyleLightContent;
-    }
-  }
-#endif
-  // it is the only non-default style available for iOS < 13
-  if (statusBarStyle == RNSStatusBarStyleLight) {
-    return UIStatusBarStyleLightContent;
-  }
-  return UIStatusBarStyleDefault;
-}
-#endif
-
-#if !TARGET_OS_TV
-+ (UIInterfaceOrientation)defaultOrientationForOrientationMask:(UIInterfaceOrientationMask)orientationMask
-{
-  if (UIInterfaceOrientationMaskPortrait & orientationMask) {
-    return UIInterfaceOrientationPortrait;
-  } else if (UIInterfaceOrientationMaskLandscapeLeft & orientationMask) {
-    return UIInterfaceOrientationLandscapeLeft;
-  } else if (UIInterfaceOrientationMaskLandscapeRight & orientationMask) {
-    return UIInterfaceOrientationLandscapeRight;
-  } else if (UIInterfaceOrientationMaskPortraitUpsideDown & orientationMask) {
-    return UIInterfaceOrientationPortraitUpsideDown;
-  }
-  return UIInterfaceOrientationUnknown;
-}
-
-+ (UIInterfaceOrientation)interfaceOrientationFromDeviceOrientation:(UIDeviceOrientation)deviceOrientation
-{
-   switch (deviceOrientation) {
-     case UIDeviceOrientationPortrait:
-       return UIInterfaceOrientationPortrait;
-     case UIDeviceOrientationPortraitUpsideDown:
-       return UIInterfaceOrientationPortraitUpsideDown;
-     // UIDevice and UIInterface landscape orientations are switched
-     case UIDeviceOrientationLandscapeLeft:
-       return UIInterfaceOrientationLandscapeRight;
-     case UIDeviceOrientationLandscapeRight:
-       return UIInterfaceOrientationLandscapeLeft;
-     default:
-       return UIInterfaceOrientationUnknown;
-   }
-}
-
-+ (UIInterfaceOrientationMask)maskFromOrientation:(UIInterfaceOrientation)orientation
-{
-  return 1 << orientation;
-}
-#endif
-
-+ (void)enforceDesiredDeviceOrientation
-{
-#if !TARGET_OS_TV
-  dispatch_async(dispatch_get_main_queue(), ^{
-    UIInterfaceOrientationMask orientationMask = UIInterfaceOrientationMaskAllButUpsideDown;
-#if defined(__IPHONE_OS_VERSION_MAX_ALLOWED) && defined(__IPHONE_13_0) && \
-    __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_13_0
-    if (@available(iOS 13, *)) {
-      UIWindow *firstWindow = [[[UIApplication sharedApplication] windows] firstObject];
-      if (firstWindow != nil) {
-        orientationMask = [firstWindow rootViewController].supportedInterfaceOrientations;
-      }
-    } else
-#endif
-    {
-      orientationMask = UIApplication.sharedApplication.keyWindow.rootViewController.supportedInterfaceOrientations;
-    }
-    UIInterfaceOrientation currentDeviceOrientation = [RNSScreenStackHeaderConfig interfaceOrientationFromDeviceOrientation:[[UIDevice currentDevice] orientation]];
-    UIInterfaceOrientation currentInterfaceOrientation = [RNSScreenStackHeaderConfig interfaceOrientation];
-    UIInterfaceOrientation newOrientation = UIInterfaceOrientationUnknown;
-    if ([RNSScreenStackHeaderConfig maskFromOrientation:currentDeviceOrientation] & orientationMask) {
-      if (!([RNSScreenStackHeaderConfig maskFromOrientation:currentInterfaceOrientation] & orientationMask)) {
-        // if the device orientation is in the mask, but interface orientation is not, we rotate to device's orientation
-        newOrientation = currentDeviceOrientation;
-      } else {
-        if (currentDeviceOrientation != currentInterfaceOrientation) {
-          // if both device orientation and interface orientation are in the mask, but in different orientations, we rotate to device's orientation
-          newOrientation = currentDeviceOrientation;
-        }
-      }
-    } else {
-      if (!([RNSScreenStackHeaderConfig maskFromOrientation:currentInterfaceOrientation] & orientationMask)) {
-        // if both device orientation and interface orientation are not in the mask, we rotate to closest available rotation from mask
-        newOrientation = [RNSScreenStackHeaderConfig defaultOrientationForOrientationMask:orientationMask];
-      } else {
-        // if the device orientation is not in the mask, but interface orientation is in the mask, do nothing
-      }
-    }
-    if (newOrientation != UIInterfaceOrientationUnknown) {
-      [[UIDevice currentDevice] setValue:@(newOrientation) forKey:@"orientation"];
-      [UIViewController attemptRotationToDeviceOrientation];
-    }
-  });
-#endif
-}
-
-+ (void)updateWindowTraits
-{
-  [RNSScreenStackHeaderConfig updateStatusBarAppearance];
-  [RNSScreenStackHeaderConfig enforceDesiredDeviceOrientation];
-}
-
-#if !TARGET_OS_TV
-// based on https://stackoverflow.com/questions/57965701/statusbarorientation-was-deprecated-in-ios-13-0-when-attempting-to-get-app-ori/61249908#61249908
-+ (UIInterfaceOrientation)interfaceOrientation
-{
-#if defined(__IPHONE_OS_VERSION_MAX_ALLOWED) && defined(__IPHONE_13_0) && \
-    __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_13_0
-  if (@available(iOS 13.0, *)) {
-    UIWindow *firstWindow = [[[UIApplication sharedApplication] windows] firstObject];
-    if (firstWindow == nil) {
-      return UIInterfaceOrientationUnknown;
-    }
-    UIWindowScene *windowScene = firstWindow.windowScene;
-    if (windowScene == nil) {
-      return UIInterfaceOrientationUnknown;
-    }
-    return windowScene.interfaceOrientation;
-  } else
-#endif
-  {
-    return UIApplication.sharedApplication.statusBarOrientation;
-  }
-}
-#endif
-
 @end
 
 @implementation RNSScreenStackHeaderConfigManager
@@ -732,16 +587,10 @@ RCT_EXPORT_VIEW_PROPERTY(largeTitleHideShadow, BOOL)
 RCT_EXPORT_VIEW_PROPERTY(hideBackButton, BOOL)
 RCT_EXPORT_VIEW_PROPERTY(hideShadow, BOOL)
 RCT_EXPORT_VIEW_PROPERTY(backButtonInCustomView, BOOL)
+RCT_EXPORT_VIEW_PROPERTY(disableBackButtonMenu, BOOL)
 // `hidden` is an UIView property, we need to use different name internally
 RCT_REMAP_VIEW_PROPERTY(hidden, hide, BOOL)
 RCT_EXPORT_VIEW_PROPERTY(translucent, BOOL)
-
-#if !TARGET_OS_TV
-RCT_EXPORT_VIEW_PROPERTY(screenOrientation, UIInterfaceOrientationMask)
-RCT_EXPORT_VIEW_PROPERTY(statusBarStyle, RNSStatusBarStyle)
-RCT_EXPORT_VIEW_PROPERTY(statusBarAnimation, UIStatusBarAnimation)
-RCT_EXPORT_VIEW_PROPERTY(statusBarHidden, BOOL)
-#endif
 
 @end
 
@@ -793,6 +642,7 @@ RCT_ENUM_CONVERTER(RNSScreenStackHeaderSubviewType, (@{
   @"right": @(RNSScreenStackHeaderSubviewTypeRight),
   @"title": @(RNSScreenStackHeaderSubviewTypeTitle),
   @"center": @(RNSScreenStackHeaderSubviewTypeCenter),
+  @"searchBar": @(RNSScreenStackHeaderSubviewTypeSearchBar),
   }), RNSScreenStackHeaderSubviewTypeTitle, integerValue)
 
 RCT_ENUM_CONVERTER(UISemanticContentAttribute, (@{
@@ -801,38 +651,6 @@ RCT_ENUM_CONVERTER(UISemanticContentAttribute, (@{
   }), UISemanticContentAttributeUnspecified, integerValue)
 
 RCT_ENUM_CONVERTER(UIBlurEffectStyle, ([self blurEffectsForIOSVersion]), UIBlurEffectStyleExtraLight, integerValue)
-
-#if !TARGET_OS_TV
-RCT_ENUM_CONVERTER(RNSStatusBarStyle, (@{
-  @"auto": @(RNSStatusBarStyleAuto),
-  @"inverted": @(RNSStatusBarStyleInverted),
-  @"light": @(RNSStatusBarStyleLight),
-  @"dark": @(RNSStatusBarStyleDark),
-  }), RNSStatusBarStyleAuto, integerValue)
-
-+ (UIInterfaceOrientationMask)UIInterfaceOrientationMask:(id)json
-{
-  json = [self NSString:json];
-  if ([json isEqualToString:@"default"]) {
-    return UIInterfaceOrientationMaskAllButUpsideDown;
-  } else if ([json isEqualToString:@"all"]) {
-    return UIInterfaceOrientationMaskAll;
-  } else if ([json isEqualToString:@"portrait"]) {
-    return UIInterfaceOrientationMaskPortrait | UIInterfaceOrientationMaskPortraitUpsideDown;
-  } else if ([json isEqualToString:@"portrait_up"]) {
-    return UIInterfaceOrientationMaskPortrait;
-  } else if ([json isEqualToString:@"portrait_down"]) {
-    return UIInterfaceOrientationMaskPortraitUpsideDown;
-  } else if ([json isEqualToString:@"landscape"]) {
-    return UIInterfaceOrientationMaskLandscape;
-  } else if ([json isEqualToString:@"landscape_left"]) {
-    return UIInterfaceOrientationMaskLandscapeLeft;
-  } else if ([json isEqualToString:@"landscape_right"]) {
-    return UIInterfaceOrientationMaskLandscapeRight;
-  }
-  return UIInterfaceOrientationMaskAllButUpsideDown;
-}
-#endif
 
 @end