+++ CHANGELOG.md
@@ -8,7 +8,11 @@ Package-specific changes not released in any SDK will be added here just before 
 ### 📚 3rd party library updates
 
 - Updated `lottie-react-native` from `3.5.0` to `4.0.2`. ([#13151](https://github.com/expo/expo/pull/13151) by [@tsapeta](https://github.com/tsapeta))
+- Updated `@react-native-picker/picker` from `1.9.11` to `1.16.1`. ([#13152](https://github.com/expo/expo/pull/13152) by [@ajsmth](https://github.com/ajsmth))
 - Updated `@react-native-community/datetimepicker` from `3.2.0` to `3.5.2`. ([#13162](https://github.com/expo/expo/pull/13162) by [@cruzach](https://github.com/13162))
+- Updated `react-native-screens` from `3.0.0` to `3.3.0`. ([#13159](https://github.com/expo/expo/pull/13159) by [@tsapeta](https://github.com/tsapeta))
+- Updated `react-native-reanimated` from `2.1.0` to `2.2.0`. ([#13161](https://github.com/expo/expo/pull/13161) by [@lukmccall](https://github.com/lukmccall))
+- Updated `react-native-svg` from `12.1.0` to `12.1.1`. ([#13154](https://github.com/expo/expo/pull/13154) by [@brentvatne](https://github.com/brentvatne))
 
 ### 🛠 Breaking changes